+++ nilearn/plotting/tests/test_img_plotting.py
@@ -398,6 +398,9 @@ def test_plot_connectome():
     # edge_threshold is too high
     plot_connectome(*args, edge_threshold=1e12)
 
+    # with colorbar=True
+    plot_connectome(*args, colorbar=True)
+
 
 def test_plot_connectome_exceptions():
     node_coords = np.arange(2 * 3).reshape((2, 3))